feat: プロジェクト構造の最適化とビルドエラーの修正
- `components` ディレクトリを feature-based に再編成 (`layout`, `features`, `common`)
- `app/page.tsx` のロジックを `hooks` に抽出し、コンポーネントを分割
- `app/layout.tsx` に `metadataBase` を追加し、Open Graph 画像の解決エラーを修正
- `BackgroundGlobe` と `LoadScreen` を `ssr: false` で動的インポートに変更し、ビルド時の SSR エラーを回避
- 不要な一時ファイルを削除
