+++ app/layout.tsx
@@ -5,8 +5,30 @@ import './globals.css'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
+  metadataBase: new URL('https://portfolio-thomas0124.vercel.app/'),
   title: 'thomas’s portfolio',
-  description: 'トーマスのポートフォリオ'
+  description: 'トーマスのポートフォリオ',
+  openGraph: {
+    type: 'website',
+    title: "thomas's portfolio",
+    description: 'トーマスのポートフォリオ',
+    url: 'https://portfolio-thomas0124.vercel.app/',
+    siteName: "thomas's portfolio",
+    images: [
+      {
+        url: 'https://portfolio-thomas0124.vercel.app/opengraph-image.png',
+        width: 1200,
+        height: 630,
+      },
+    ],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    site: '@Tomas_engineer',
+    title: "thomas's portfolio",
+    description: 'トーマスのポートフォリオ',
+    images: ['https://portfolio-thomas0124.vercel.app/twitter-image.png'],
+  },
 }
 
 export default function RootLayout({
@@ -16,21 +38,6 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ja">
-      <meta property="og:type" content="website" />
-      <meta property="og:title" content="thomas's portfolio" />
-      <meta property="og:description" content="トーマスのポートフォリオ" />
-      <meta property="og:url" content="https://portfolio-thomas0124.vercel.app/" />
-      <meta property="og:site_name" content="thomas's portfolio" />
-      <meta property="og:image" content="https://portfolio-thomas0124.vercel.app/opengraph-image.png" />
-      <meta property="og:image:type" content="image/png" />
-      <meta property="og:image:width" content="1200" />
-      <meta property="og:image:height" content="630" />
-
-      <meta property="twitter:title" content="thomas's portfolio" />
-      <meta property="twitter:description" content="トーマスのポートフォリオ" />
-      <meta property="twitter:image" content="https://portfolio-thomas0124.vercel.app/twitter-image.png" />
-      <meta name="twitter:card" content="summary_large_image" />
-      <meta name="twitter:site" content="@Tomas_engineer" />
       <body className={inter.className}>{children}</body>
     </html>
   )
